Move toast logic into ToastManager
