+++ react/components/Toast/index.js
@@ -1,7 +1,6 @@
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
-import Toast from './Toast'
-import isString from 'lodash/isString'
+import ToastManager from './ToastManager'
 
 const ToastContext = React.createContext({
   showToast: () => {},
@@ -12,103 +11,25 @@ class ToastProvider extends Component {
   constructor(props) {
     super(props)
 
-    this.container = React.createRef()
+    this.toastManager = React.createRef()
   }
 
-  state = {
-    currentToast: null,
-    nextToast: null,
-    isToastVisible: true,
-  }
-
-  showToast = args => {
-    if (isString(args)) {
-      args = { message: args }
-    }
-    const { message = '', action, duration } = args
-
-    if (this.state.currentToast) {
-      // If there is a toast present already, queue up the next toast
-      // It will be displayed when the current toast is closed, on handleToastClose
-      this.setState({
-        nextToast: {
-          message,
-          action,
-          duration,
-        },
-      })
-      this.hideToast()
-    } else {
-      this.setState({
-        currentToast: {
-          message,
-          action,
-          duration,
-        },
-        isToastVisible: true,
-      })
-    }
+  showToast = message => {
+    this.toastManager &&
+      this.toastManager.current &&
+      this.toastManager.current.showToast &&
+      this.toastManager.current.showToast(message)
   }
 
   hideToast = () => {
-    this.setState({
-      isToastVisible: false,
-    })
-  }
-
-  handleToastClose = () => {
-    this.setState(state => {
-      return {
-        // If there is a toast queued up, shows it.
-        // Otherwise, nextToast will be null, and state.toast will be cleared up
-        currentToast: state.nextToast,
-        isToastVisible: !!state.nextToast,
-        nextToast: null,
-      }
-    })
-  }
-
-  getParentBounds = () => {
-    const parentContainer =
-      this.container.current && this.container.current.parentNode
-    return (
-      parentContainer &&
-      parentContainer.getBoundingClientRect &&
-      parentContainer.getBoundingClientRect()
-    )
-  }
-
-  updateContainerBounds = () => {
-    const windowBounds = {
-      left: 0,
-      right: window.innerWidth,
-      top: 0,
-      bottom: window.innerHeight,
-    }
-
-    const bounds =
-      (this.props.positioning === 'parent' && this.getParentBounds()) ||
-      windowBounds
-
-    if (this.container.current) {
-      this.container.current.style.left = `${bounds.left}px`
-      this.container.current.style.right = `${window.innerWidth -
-        bounds.right}px`
-      this.container.current.style.top = `${bounds.top}px`
-      this.container.current.style.bottom = `${Math.max(
-        0,
-        window.innerHeight - bounds.bottom
-      )}px`
-    }
-  }
-
-  componentDidUpdate() {
-    this.updateContainerBounds()
+    this.toastManager &&
+      this.toastManager.current &&
+      this.toastManager.current.hideToast &&
+      this.toastManager.current.hideToast()
   }
 
   render() {
-    const { currentToast } = this.state
-    const { children } = this.props
+    const { children, positioning } = this.props
     return (
       <ToastContext.Provider
         value={{
@@ -116,22 +37,7 @@ class ToastProvider extends Component {
           hideToast: this.hideToast,
         }}>
         {children}
-        <div
-          className="fixed z-5 overflow-hidden"
-          ref={this.container}
-          style={{
-            pointerEvents: 'none',
-          }}>
-          {currentToast && (
-            <Toast
-              message={currentToast.message}
-              action={currentToast.action}
-              duration={currentToast.duration}
-              visible={this.state.isToastVisible}
-              onClose={this.handleToastClose}
-            />
-          )}
-        </div>
+        <ToastManager positioning={positioning} ref={this.toastManager} />
       </ToastContext.Provider>
     )
   }
